When removing old refresh and subscribe job states, only access the DB to read them if there actually any records to delete.
Before this commit, DestroyOldJobStatesJob.perform always executed two queries for each of subscribe_feed_job_states and refresh_feed_job_states. The first query counted how many records were to be deleted, and the second query selected them to actually delete them.

If there are no records to delete (the count is zero), it is wasteful to perform a SELECT which returns zero records anyway. A couple of if conditionals have been added so that the second query is only executed if the count of records to delete is greater than zero.

+++ app/jobs/destroy_old_job_states_job.rb
@@ -16,11 +16,13 @@ class DestroyOldJobStatesJob
     Rails.logger.info "Destroying job states created before #{time_old}"
 
     old_refresh_feed_jobs = RefreshFeedJobState.where 'created_at < ?', time_old
-    Rails.logger.info "Destroying #{old_refresh_feed_jobs.count} old instances of RefreshFeedJobState"
-    old_refresh_feed_jobs.destroy_all
+    old_refresh_feed_jobs_count = old_refresh_feed_jobs.count
+    Rails.logger.info "Destroying #{old_refresh_feed_jobs_count} old instances of RefreshFeedJobState"
+    old_refresh_feed_jobs.destroy_all if old_refresh_feed_jobs_count > 0
 
     old_subscribe_jobs = SubscribeJobState.where 'created_at < ?', time_old
-    Rails.logger.info "Destroying #{old_subscribe_jobs.count} old instances of SubscribeJobState"
-    old_subscribe_jobs.destroy_all
+    old_subscribe_jobs_count = old_subscribe_jobs.count
+    Rails.logger.info "Destroying #{old_subscribe_jobs_count} old instances of SubscribeJobState"
+    old_subscribe_jobs.destroy_all if old_subscribe_jobs_count > 0
   end
 end
